MMIOT-103-1 imx8: optee: fix build issue after upgrading to optee3.0
This patch fixes below error when building under Android

    external/optee_test/host/xtest/xtest_main.c:105: error: undefined
    reference to 'install_ta_runner_cmd_parser'
    clang.real: error: linker command failed with exit code 1 (use -v
    to see invocation)

The CFG_SECSTOR_TA_MGMT_PTA variable must be enabled by default,
otherwise the install_ta.c source file is not compiled, while the code
is refering some functions from it.
Indeed, the flag is set in the conf.h generated by optee_os. Thus, when
code is compiled, install_ta functions are used in other part of the
code. However, the conf.mk didn't exist when the optee_test Android.mk
file was parsed. So install_ta.c is not in the srcs files.
This is causing linker issue at the end.

Change-Id: I85f159091eea0715b5554fd67b08bd72d8d1c1cd
Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/4271
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>