MA-22343 fix boot up failure when update gms to 202404
when the gms package is updated to the version of 202404, there is still
build error related to Launcher3, removed the integration of
SearchLauncherQuickStep to avoid the issue. when boot up, there is
failure log related to priapp permissions. Change the permissions in
this commit to eleminate the failure.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I3da108e3a656e5ad069b3041998a33a38b7eb61c
Reviewed-on: http://androidsource.nxp.com/project/24492
Reviewed-by: Elven Wang <elven.wang@nxp.com>
(cherry picked from commit 373d56f25234fc50f7236b4a0999f9e708499c81)
Reviewed-on: http://androidsource.nxp.com/project/24498