MA-22686 Adjust the default swiotlb pool size
With the enablement of dynamic SWIOTLB (CONFIG_SWIOTLB_DYNAMIC),
the swiotlb pool would be allocated dynamically when the existing
pool size can't be satisfied.

Instead of allocating a large pool in the beginning, this commit
maintains a minimum required swiotlb pool and let the kernel
allocate more if it can't satisfy the incoming request. This will
help save some memory.

The formula to calculate the swiotlb pool size is:
    swiotlb = 128 * num_cpus

For imx8mp and imx95, use the default swiotlb pool size (64MB).

Change-Id: I7c7ceaceff0b007399bb5e32d76e4351d4f70ca4
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25432
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>