MA-22777 inherit quarterly release config to create one named nxp_stable
With trunk stable feature, a release config need to be specified when
lunch a target.

The quarterly release config name have 4 fields, for example ap3a.
  a: represents the year. Starting with A=2024
  p: represents the product type. p=platform, t=tv, w=wearOS...
  3: means release of the year. 1-9a-z for 35 max.
  a: means release type. 1-9 for pre-release, a for primary release. b
     is reserver

it is not convenient to notice that a new quarterly release config is
avaiable for use and it should be specified when lunch build targets.

Create a config named nxp_stable, it keeps inheriting the current
quarterly release config. then is can always be used when lunch a build
target.

Change-Id: Iacbac6b6cbd0e043563f3b3fdf9a20999230416c
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25706
Reviewed-by: Elven Wang <elven.wang@nxp.com>
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>