MA-22648-1 integrate the privapp permission file for EmergencyInfo
The EmergencyInfo app was removed in the patch to remove telephony
things.

When GMS is integrated, the EmergencyInfoGms is integrated because of
the PersonalSafety app. This EmergencyInfoGms overrides the
EmergencyInfo module, and it needs the permission file of this module or
the system won't boot up.

Integrate the permission file individually to solve this boot issue.

Change-Id: Idb3dab8cc91bf6eaff0639988f48003757d81367
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26117
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>