MA-23015 audio: Fix cts AudioRecordSharedAudioTest#testCapturesMatch
The case starts two AudioRecorder and expects the data
recorded by them are identical, format is 16000, mono.
But evk_8ulp and mek_8q doesn't support the format, framework
will uses the mixer to convert the format. This may lead
the mismatch issue.
Update the policy file to support more rates.

Note, 8ulp and 8q may use the same card(wm8960) to record and playback
at the same time, this requires the same audio in/out format
because they share the same clock.

Change-Id: I32c6643aeef99ec9997f871f3da425d4541a2b10
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26544
Reviewed-by: Hui Fang <hui.fang@nxp.com>