MA-23160 audio: Move compressed offload port to sof specific policy for mek_8q
The card cs42888 has been set as line-out device. With
compressed offload port routed to it, some CTS cases
will try to test it and fail due to sof card not present.
  android.media.audio.cts.AudioTrackOffloadTest CtsMediaAudioTestCases

Move the compressed port out of the default policy file.

To support sof, do one more steps to override the policy:
 # remount
 # cp /vendor/etc/audio_policy_configuration_sof.xml /vendor/vendor_overlay_soc/imx8qm/vendor/etc/audio_policy_configuration.xml
or
 # cp /vendor/etc/audio_policy_configuration_sof.xml /vendor/vendor_overlay_soc/imx8qxp/vendor/etc/audio_policy_configuration.xml
 # sync && reboot

Change-Id: I5908ddbe4f690eb4dab9081e5ef1e80a3b7296d8
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26612
Reviewed-by: Hui Fang <hui.fang@nxp.com>