MA-23132 align usb tethering behavior between android14 and 15
In the repo packages/modules/Connectivity, There is a commit with the
following header message.

    Include both eth%d and usb%d ethernet interfaces on V+

With that commit, from android15, both eth* and usb* are taken as
tethering upstream by default. As usb* can represent both a usb ethernet
dongle and a usb tethering downstream, this causes usb tethering restart
when try to use it, and tethering cannot work as expected.

In this patch, override the config_ethernet_iface_regex value so that
usb* won't be used as tethering upstream, this is the same behavior as
android14.

Change-Id: I618737ce17c7745902838751e6af0a73fb707cfc
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26637
Reviewed-by: Elven Wang <elven.wang@nxp.com>