MA-22245 set dwc3 device autosuspend delay to be 0 in powersave image
In imx kernel, a patch is used to disable the use of dwc3 autosuspend.
As there are some patches not upstreamed, this patch does not take
effect with GKI.

It is not able to disable the use of dwc3 autosuspend with sysfs, but
the autosuspend delay can be set to be zero with sysfs, the same effect
can be achieved.

With this patch applied, powersave image with GKI can dectect usb device
plug out and then plug in in 5 seconds. Also, the issue in ticket
MA-23224 can also be avoided.

Change-Id: I4d5384fb2d9e3d30d9515871b3f02eac87aa7aec
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27402
Reviewed-by: Elven Wang <elven.wang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27874
Reviewed-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>