MA-23189-5 move partition table image generation out of soong buid
In Android16, HOST python tools are not allowed to be used, and soong
build does not support python2. the bpttool uses python2 and not
maintained by Google.

In this patch, move the partition table image generate process which
uses bpttool out of the soong build, so HOST python2 can be used.

Change-Id: I2bf48fc452bca3f818a06933c63bd633c62249bc
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27453
Reviewed-by: Elven Wang <elven.wang@nxp.com>