MA-23395-1 tune boot time IO to optimze boot time
iostats: used to control (on/off) the iostats accounting of the
    disk. turn off the accounting.

scheduler: used to show and set the IO scheduler of the current device.
    Default to use mq-deadline. Use kyber in boot time to get better
    boot time performance.

nr_requests: how many requests may be allocated in the block layer for
    read or write requests. 128 by default. set it to be 256 in boot
    time.

read_ahead_kb: Maximum number of kilobytes to read-ahead for filesystems
    on this block device. 128 by default. set it to be 1024 during boot
    time.

if dm-verity is not disabled, dm-0 to dm-5 are dm-linear devices, dm-6
to dm-11 are dm-verity devices currently, dm-6 to dm-11 are mounted and
directly read data from.

TEST: on evk_95, it saves about 0.53s boot time.

Change-Id: Ieb4c2d7b4e238695e7acaca7304dae6be6feb509
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/28087
Reviewed-by: Elven Wang <elven.wang@nxp.com>
Reviewed-by: Ji Luo <ji.luo@nxp.com>