MA-23500 audio: MQS: Can't record audio from J23 headphone port
The MQS dts only has two cards: mqs for output and wm8962 for input
and output. When headphone is connected to wm8962, the card is
used for both input and output. We rely on the audio hal to
do the sample rates conversion, it only supports 48000/2c/16bit.

Remove the mono channel in policy as the resample doesn't support it.

Change-Id: I392246275d6eadc7a33d25a46529215551b2d1c9
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27643
Reviewed-by: Hui Fang <hui.fang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27746