MA-23189-6 remove selinux lable for /dev/sda* to fix build errors
After the codebase upgrade, there are build errors related to the
selinux label of /dev/sda*. For the label used for files under /dev/, it
should be associated with the "dev_type" attribute, but "fuse" is not
associated with "dev_type".

As there is no /dev/sda* files now. remove the selinux label.

Change-Id: I2cb988ad2ce0cf36b59367569d49cde5a7d37a9f
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27866
Reviewed-by: Elven Wang <elven.wang@nxp.com>