MA-23189-6 move partition table image generation out of soong buid on 943
In Android16, HOST python tools are not allowed to be used, and soong
build does not support python2. the bpttool uses python2 and not
maintained by Google.

In this patch, move the partition table image generate process which
uses bpttool out of the soong build, so HOST python2 can be used.

Signed-off-by: Richard Liu <xuegang.liu@nxp.com>
Change-Id: I6ef2a5550c4cd363ee13bfae7076f1e9c7fa642b
Reviewed-on: http://androidsource.nxp.com/project/28023
Reviewed-by: Faqiang Zhu <faqiang.zhu@nxp.com>