MA-23675-2 imx95: fix vpu_firmware_load fail
The vpu_firmware_load will trap to TEE and call some generic
crypto APIs from ELE to do the key verification. However, the
generic crypto APIs are moved to a run-time ELE FW now and
the FW is not loaded when vpu_firmware_load starts.

The ELE runtime FW load was triggered by calling the ELE
driver from the non-secure usersapce, which is a little tricky.

This commit moves the nvmd to an earlier stage so it can trigger
the run-time FW load before starting vpu_firmware_load. The
vpu_firmware_load was moved to 'zygote-start' to avoid random
ELE panic because of FW loading and ELE requests from the TEE
happen at the same time.

Change-Id: I6108fb12ee2c09d8c968183271d19de8a8b02a21
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/28428
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>