MA-23596-1 increase a backup timeout value to pass a cts test on 8qxp
With GMS integrated, the following case fails with a quite high
possibility:

 module: CtsBackupHostTestCases
 case: android.cts.backup.SyncAdapterSettingsHostSideTest#testIsSyncableChanged_ifTurnOnSyncEnabled

This case execute the shell command "bmgr backupnow android" and check
the output, it fails if the output has nothing related to "android"
backup successfully. This is related to the kv backup agent timeout
value, the default value is 30s, set it to be 60s, so this case can
always pass the test.

Change-Id: I84a382340f5753f0ad79e2dd3491fa8f0802faa7
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/28976
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
(cherry picked from commit 516af8052f200296a54ca431b3c6888093f0f5fb)
Reviewed-on: http://androidsource.nxp.com/project/29693