MA-24100 Only include system_dlkm staging modules when LOADABLE_KERNEL_MODULE is true
When building the kernel with imx_v8_android_defconfig, the following
warning will appear during boot: disagrees about version of symbol xxxxxx

This is because the imx943 requires CONFIG_NET_DSA=y and
CONFIG_NET_DSA_NETC_SWITCH=y. In addition, imx_v8_android_defconfig does
not include the GKI bypassing CONFIG, such as CONFIG_IMX_GKI_8Q_FIX=y
, # CONFIG_PAGE_POOL_STATS is not set and CONFIG_IMX_GKI_FIX=y.

As a result, there is a KMI break between the modules in system_dlkm and
the kernel, which leads to the warning. To avoid this issue, use
LOADABLE_KERNEL_MODULE for control.

Change-Id: I3975eafdaa16bc5bb79deeb8054810a290beb2a0
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29790
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>