MA-23395-3 use prefetch to save boot time
Google provided a mechanism to prefetch pages from read-only partitions
to optimize the boot time.

Refer to the code under system/core/init/libprefetch/prefetch/ to get
knowledge of prefetch record and replay.

Set the property ro.prefetch_boot.enabled=true to enable this mechanism.
Other parameters can be adapted are listed below, categorized into 2
part:
1. record related paramters
  1) timepoint to start the prefetch_record service
  2) the duration of the record. this can be set with 2 method
    a. directly set the duration with prop ro.prefetch_boot.duration_s
    b. set the prefetch_boot.record_stop at a point to stop the replay

2. replay related parameters
  1) timepoint to start the prefetch_replay service
  2) Number of threads required to replay, set with prop
     ro.prefetch_boot.io_depth. It has a default value.
  3) The maximum replay fd count, set with the prop
     ro.prefetch_boot.max_fds. it also has a default value.

This parameters may need to be set with different values for the same
SoC on different board, So the settings in this patch are references.
Customers may adapt the values to get the best result.

Tried on evk_8ulp, evk_8mm and evk_95, adjusted the parameters and then
test the boot time, got the following time saving result:

  evk_8ulp: 1.6s
  evk_8mm: 0.62s
  evk_95: 0.74s

The parameters may need to be adjusted to get the best result when the
codebase is upgraded.

Change-Id: I025404938867d22bf2bf7adf0c4aaa03820665ec
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29816
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>