MA-24129 check images used by uuu exist or not
if a file to be used by uuu does not exist, uuu terminated with log as
below. one need to check which image does not exist

  Wait for Known USB Device Appear...
  Error: FB: image file not found

In this patch, check whether a file exist, print clear log and exit early
if a file does not exist

For the linux version uuu script, register the clean function to be
executed at the exit of the script.

Change-Id: Ia9d920e31c9c76ca2eeaa2774c7ae6680edcc3fe
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29823
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>