MA-24049-6 Include dtbs to vendor_boot image
According to the GBL requirement, the base device tree can come
from boot, vendor_boot or dtb partition, the overlay device tree
can come from dtbo partion. Only one base device tree can be selected,
one or more overlay device tree can be applied to the base device
tree.

We didn't enable the device tree overlay but the device tree was
compiled into the dtbo partition, which violates the GBL's requirement.

This commit moves the dtbs to vendor_boot image, the dtbs would be
organized in dt_table_entry format. Each dtb would be assigned with
an "id" in the sequence as it appears in TARGET_BOARD_DTS_CONFIG
variable. The same id and dtb name mapping would also be maintained
in u-boot, the name and sequence should be kept exact same.

Currently this option is only enabled when SUPPORT_GBL is set to true.

Change-Id: I4cd0ce20d747ff1ac55ccb2fa3a92d90cacc4927
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29909
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>